refactor(storage): move space analyzer into module
