+++ app/index.ts
@@ -21,6 +21,7 @@ import IGeesomeAccountStorageModule from "./modules/accountStorage/interface.js"
 import IGeesomeCommunicatorModule from "./modules/communicator/interface.js";
 import IGeesomeStaticIdModule from "./modules/staticId/interface.js";
 import IGeesomeStorageModule from "./modules/storage/interface.js";
+import IGeesomeStorageSpaceModule from "./modules/storageSpace/interface.js";
 import IGeesomeDriversModule from "./modules/drivers/interface.js";
 import IGeesomeContentModule from "./modules/content/interface.js";
 import IGeesomeInviteModule from "./modules/invite/interface.js";
@@ -118,6 +119,7 @@ function getModule(config, appPass) {
       accountStorage: IGeesomeAccountStorageModule,
       communicator: IGeesomeCommunicatorModule,
       storage: IGeesomeStorageModule,
+      storageSpace: IGeesomeStorageSpaceModule,
       entityJsonManifest: IGeesomeEntityJsonManifestModule
     };
