+++ app/interface.ts
@@ -13,6 +13,7 @@ import IGeesomeAsyncOperationModule from "./modules/asyncOperation/interface.js"
 import IGeesomeCommunicatorModule from "./modules/communicator/interface.js";
 import IGeesomeStaticIdModule from "./modules/staticId/interface.js";
 import IGeesomeStorageModule from "./modules/storage/interface.js";
+import IGeesomeStorageSpaceModule from "./modules/storageSpace/interface.js";
 import IGeesomeDriversModule from "./modules/drivers/interface.js";
 import IGeesomeContentModule from "./modules/content/interface.js";
 import IGeesomeInviteModule from "./modules/invite/interface.js";
@@ -44,6 +45,7 @@ export interface IGeesomeApp {
     group: IGeesomeGroupModule;
     accountStorage: IGeesomeAccountStorageModule;
     storage: IGeesomeStorageModule;
+    storageSpace: IGeesomeStorageSpaceModule;
     communicator: IGeesomeCommunicatorModule;
     entityJsonManifest: IGeesomeEntityJsonManifestModule;
     database: IGeesomeDatabaseModule;
