+++ docs/database-scalability-inventory.md
@@ -61,7 +61,7 @@ This file is generated by `npm run database:scalability:update`. Do not hand-edi
 | Category feeds | getCategoryPosts | joins group/category pivots only for page post ID selection, supports (publishedAt,id) cursor pages and includeTotal=false offset pages, then hydrates contents/group for the bounded page | cursor pages and explicit includeTotal=false offset pages skip counts; legacy offset callers still pay count/offset cost by default | `app/modules/groupCategory/index.ts` |
 | Content preview serving | getContentByStorageId(findByPreviews) | legacy/global manifest lookups still use deterministic content helpers, ownerless imports can only reuse existing shared rows, public metadata is projected, content writes maintain a storageObject registry row, and shared storage/header reads prefer storageObject metadata with Content fallback | shared reads are stable across duplicate user-owned rows, no longer create new ownerless library rows, public metadata hides private DB rows, and shared serving now prefers canonical physical metadata while preserving old-row fallback | `app/modules/database/index.ts` |
 | Content deletion | deleteFileCatalogItem deleteContent | destroys catalog item first; only destroys content/physical storage after DB reference checks, exposing content/storage blocker reasons for future delayed-GC or operator reporting; checks include all known derived storage columns/current static-ID refs plus successful remote pins marked on StorageObject.isPinned and Content.isPinned | same-storage, preview, known DB-visible derived storage columns, current static-ID refs, and canonical local pin state are covered; IPFS DAG child refs inside generated output, remote pin reconciliation, and async garbage collection still need a fuller lifecycle | `app/modules/database/index.ts + app/modules/fileCatalog/index.ts` |
-| Storage space analysis | storage analyzer aggregate helpers | read-only helpers, AdminRead API routes, model-sync cached snapshots, file-catalog folder and group-post drilldowns, and asyncOperation-owned refresh queue jobs expose overview totals, MIME/type breakdowns, largest content rows, largest catalog files/folders, largest groups, and largest published posts while separating logical content bytes from deduplicated physical storage bytes | backend aggregate, API, cached snapshot, generic async refresh queue, file-catalog folder drilldown, and first group-post drilldown seams are present; generated-output DAG accounting, restored-data query evidence, and finer per-query progress remain | `app/modules/database/storageSpaceUsageHelpers.ts + app/modules/api/api.ts + app/modules/database/models/storageSpaceSnapshot.ts` |
+| Storage space analysis | storage analyzer aggregate helpers | read-only helpers, AdminRead API routes, model-sync cached snapshots, file-catalog folder and group-post drilldowns, and asyncOperation-owned refresh queue jobs expose overview totals, MIME/type breakdowns, largest content rows, largest catalog files/folders, largest groups, and largest published posts while separating logical content bytes from deduplicated physical storage bytes | backend aggregate, API, cached snapshot, generic async refresh queue, file-catalog folder drilldown, and first group-post drilldown seams are present; generated-output DAG accounting, restored-data query evidence, and finer per-query progress remain | `app/modules/storageSpace/queryHelpers.ts + app/modules/storageSpace/api.ts + app/modules/database/models/storageSpaceSnapshot.ts` |
 | File catalog publish | publishFolder / makeFolderChildrenStorageDirsAndCopyFiles | recursively scans child folders and files in id-ordered batches before copying storage files | large folder publishes avoid the public browsing cap and avoid one unbounded child read per folder | `app/modules/fileCatalog/index.ts` |
 | Quota checks | checkFileSizeAndSaveByStream | quota sum before upload stream, then content row and upload action commit under a locked UserLimit re-check | visible content/action commits are serialized against the active limit; bytes can still be streamed before final commit rejection, so true pre-stream reservation remains future work | `app/modules/content/index.ts` |
 | Async operation ledger | getUserAsyncOperationList / closeAllAsyncOperation / cleanupFinishedAsyncOperations | lists by user/inProcess/name ordered by createdAt, closes in-process rows on startup, and deletes old finished rows in a bounded batch | lookup, startup sweep, queue lifecycle, and finished-row retention cleanup are present; operation history beyond the latest window is intentionally temporary | `app/modules/asyncOperation/index.ts` |
