+++ docs/security-route-inventory.md
@@ -11,7 +11,7 @@ This file is generated by `npm run security:route-inventory:update`. Do not hand
 - API route registrations: 150
 - Public registrations: 27
 - Authorized registrations: 123
-- Authorized registrations without nearby core-permission checks: 104
+- Authorized registrations without nearby core-permission checks: 105
 
 Interpretation notes:
 
@@ -35,23 +35,13 @@ Interpretation notes:
 | api | POST | `/v1/admin/permissions/core/remove_permission` | authorized | AdminSetPermissions | - | `app/modules/api/api.ts` |
 | api | POST | `/v1/admin/permissions/core/set_permissions` | authorized | AdminSetPermissions | - | `app/modules/api/api.ts` |
 | api | POST | `/v1/admin/set-user-limit` | authorized | - | token only; module/user ownership checks expected | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/file-catalog-folders` | authorized | AdminRead | content/storage boundary | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/group-posts` | authorized | AdminRead | - | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/overview` | authorized | AdminRead | - | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/snapshot` | authorized | AdminRead | - | `app/modules/api/api.ts` |
-| api | POST | `/v1/admin/storage-space/snapshot/refresh-async` | authorized | AdminRead | - | `app/modules/api/api.ts` |
-| api | POST | `/v1/admin/storage-space/snapshot/refresh` | authorized | AdminRead | - | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/top-contents` | authorized | AdminRead | content/storage boundary | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/top-file-catalog-items` | authorized | AdminRead | content/storage boundary | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/top-groups` | authorized | AdminRead | - | `app/modules/api/api.ts` |
-| api | GET | `/v1/admin/storage-space/type-breakdown` | authorized | AdminRead | - | `app/modules/api/api.ts` |
 | api | GET | `/v1/api/v0/refs*` | public | - | public entrypoint | `app/modules/api/api.ts` |
 | api | POST | `/v1/get-user-by-api-token` | authorized | - | token only; module/user ownership checks expected | `app/modules/api/api.ts` |
 | api | GET | `/v1/ipld/*` | public | - | public entrypoint | `app/modules/api/api.ts` |
 | api | GET | `/v1/is-empty` | public | - | public entrypoint | `app/modules/api/api.ts` |
 | api | POST | `/v1/login/password` | public | - | authentication flow; public entrypoint | `app/modules/api/api.ts` |
 | api | GET | `/v1/node-address-list` | public | - | public entrypoint | `app/modules/api/api.ts` |
-| api | POST | `/v1/save-object` | authorized | AdminRead | - | `app/modules/api/api.ts` |
+| api | POST | `/v1/save-object` | authorized | - | token only; module/user ownership checks expected | `app/modules/api/api.ts` |
 | api | POST | `/v1/setup` | public | - | public entrypoint | `app/modules/api/api.ts` |
 | api | GET | `/v1/user` | authorized | - | token only; module/user ownership checks expected | `app/modules/api/api.ts` |
 | api | GET | `/v1/user/api-key-list` | authorized | - | token only; module/user ownership checks expected | `app/modules/api/api.ts` |
@@ -158,6 +148,16 @@ Interpretation notes:
 | staticId | HEAD | `/ipns/*` | public | - | public entrypoint | `app/modules/staticId/api.ts` |
 | staticId | GET | `/v1/resolve/:storageId` | public | - | public entrypoint | `app/modules/staticId/api.ts` |
 | staticId | GET | `/v1/self-account-id` | public | - | public entrypoint | `app/modules/staticId/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/file-catalog-folders` | authorized | AdminRead | content/storage boundary | `app/modules/storageSpace/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/group-posts` | authorized | AdminRead | - | `app/modules/storageSpace/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/overview` | authorized | AdminRead | - | `app/modules/storageSpace/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/snapshot` | authorized | AdminRead | - | `app/modules/storageSpace/api.ts` |
+| storageSpace | POST | `/v1/admin/storage-space/snapshot/refresh-async` | authorized | AdminRead | - | `app/modules/storageSpace/api.ts` |
+| storageSpace | POST | `/v1/admin/storage-space/snapshot/refresh` | authorized | AdminRead | - | `app/modules/storageSpace/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/top-contents` | authorized | AdminRead | content/storage boundary | `app/modules/storageSpace/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/top-file-catalog-items` | authorized | AdminRead | content/storage boundary | `app/modules/storageSpace/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/top-groups` | authorized | AdminRead | - | `app/modules/storageSpace/api.ts` |
+| storageSpace | GET | `/v1/admin/storage-space/type-breakdown` | authorized | AdminRead | - | `app/modules/storageSpace/api.ts` |
 | telegramClient | POST | `/v1/soc-net/telegram/channel-info` | authorized | - | social account credential boundary; token only; module/user ownership checks expected | `app/modules/telegramClient/api.ts` |
 | telegramClient | POST | `/v1/soc-net/telegram/channels` | authorized | - | social account credential boundary; token only; module/user ownership checks expected | `app/modules/telegramClient/api.ts` |
 | telegramClient | POST | `/v1/soc-net/telegram/login` | authorized | - | authentication flow; social account credential boundary; token only; module/user ownership checks expected | `app/modules/telegramClient/api.ts` |
