+++ docs/todo.md
@@ -17,7 +17,7 @@ Corrections and added requirements:
 - Add security review of API and encryption flows to the TODO. Status: tracked in [#782](https://github.com/galtproject/geesome-node/issues/782) and added as a fast-delivery security gate.
 - API documentation tooling should be handled through microwave-hub submodules for [`apidoc-template`](https://github.com/MicrowaveDev/apidoc-template) and [`apidoc-plugin-ts`](https://github.com/MicrowaveDev/apidoc-plugin-ts). Status: hub submodule tracking is in [Microwave Hub #2](https://github.com/MicrowaveDev/microwave-hub/issues/2), planning was tracked in [#787](https://github.com/galtproject/geesome-node/issues/787), vulnerable `apidoc-core` removal was tracked in [#802](https://github.com/galtproject/geesome-node/issues/802), final git-URL wiring was tracked in [#804](https://github.com/galtproject/geesome-node/issues/804), plugin-master repoint was tracked in [#806](https://github.com/galtproject/geesome-node/issues/806), request-body annotation support was tracked in [#808](https://github.com/galtproject/geesome-node/issues/808), all-module generation for existing annotated specs is tracked in [#810](https://github.com/galtproject/geesome-node/issues/810), practical remaining route coverage is tracked in [#812](https://github.com/galtproject/geesome-node/issues/812), and final examples/errors/render polish is tracked in [#813](https://github.com/galtproject/geesome-node/issues/813).
 - Add database scalability review for the possibility of storing hundreds of thousands of posts and their content records in groups. Status: tracked in [#880](https://github.com/galtproject/geesome-node/issues/880) and added as a dedicated review slice.
-- Add a global UI menu item and screen for storage-space analysis: show which file catalogs and groups use how much space, then let operators drill into largest files, file types, duplicate/shared storage IDs, and cleanup candidates. Status: added as a dedicated planning slice; first `geesome-node` aggregate helpers, AdminRead API routes, cached snapshot routes/table, queued async refresh path, file-catalog root/folder drilldown, group-post drilldown, and the `geesome-ui` global screen exist. Remaining implementation should coordinate generated-output DAG accounting, restored-data measurement, finer per-query progress, and deeper cleanup/detail views.
+- Add a global UI menu item and screen for storage-space analysis: show which file catalogs and groups use how much space, then let operators drill into largest files, file types, duplicate/shared storage IDs, and cleanup candidates. Status: added as a dedicated planning slice; first `geesome-node` aggregate helpers live in a dedicated `storageSpace` module with AdminRead API routes, cached snapshot routes/table, queued async refresh path, file-catalog root/folder drilldown, group-post drilldown, and the `geesome-ui` global screen. Remaining implementation should coordinate generated-output DAG accounting, restored-data measurement, finer per-query progress, and deeper cleanup/detail views.
 
 Last issue snapshot: 2026-05-03 from `galtproject/geesome-node` open GitHub issues and PRs.
 
@@ -396,7 +396,7 @@ Verification:
 
 ### 12. Storage Space Analysis UI
 
-Status: in progress, with the first backend aggregate/API seams, cached snapshot routes/table, queued async refresh path, file-catalog root/folder drilldown endpoint, group-post drilldown endpoint, and `geesome-ui` global navigation/screen in place. `database.getStorageSpaceOverview`, type breakdown, top content, top file-catalog file/folder, top group, and group-post helpers now report logical content bytes separately from deduplicated physical storage bytes; AdminRead routes expose live results under `admin/storage-space/*`; and a model-sync `StorageSpaceSnapshot` table can store timestamped analyzer snapshots for operator refreshes. The snapshot refresh now uses the shared async-operation queue runner under the `storage-space-snapshot` module for progress polling while the database module owns only the storage-space job payload and snapshot calculation. Remaining work is generated-output DAG accounting, restored-data measurement, finer per-query progress, and deeper cleanup/detail views.
+Status: in progress, with the first backend aggregate/API seams, cached snapshot routes/table, queued async refresh path, file-catalog root/folder drilldown endpoint, group-post drilldown endpoint, and `geesome-ui` global navigation/screen in place. The dedicated `storageSpace` module reports logical content bytes separately from deduplicated physical storage bytes for overview, type breakdown, top content, top file-catalog file/folder, top group, and group-post views; AdminRead routes expose live results under `admin/storage-space/*`; and a model-sync `StorageSpaceSnapshot` table can store timestamped analyzer snapshots for operator refreshes. The snapshot refresh now uses the shared async-operation queue runner under the `storage-space-snapshot` module for progress polling, while the database module keeps only the synced snapshot model and shared SQL connection. Remaining work is generated-output DAG accounting, restored-data measurement, finer per-query progress, and deeper cleanup/detail views.
 
 Goal: give node operators a global space-analysis view that explains where storage is going and what can be investigated safely before cleanup or quota decisions.
