+++ test/storageSpaceApiUnit.test.ts
@@ -1,5 +1,6 @@
 import assert from "assert";
 import registerCoreApi from "../app/modules/api/api.js";
+import registerStorageSpaceApi from "../app/modules/storageSpace/api.js";
 import {CorePermissionName} from "../app/modules/database/interface.js";
 
 function createCoreApiHarness(appOverrides: any = {}) {
@@ -23,6 +24,7 @@ function createCoreApiHarness(appOverrides: any = {}) {
 	const app = {
 		ms: {
 			database: {},
+			storageSpace: {},
 			communicator: {},
 			storage: {},
 		},
@@ -30,7 +32,9 @@ function createCoreApiHarness(appOverrides: any = {}) {
 		...appOverrides,
 	};
 
+	app.ms.api = module;
 	registerCoreApi(app as any, module as any);
+	registerStorageSpaceApi(app as any, app.ms.storageSpace as any);
 
 	return {
 		async call(method, path, req: any = {}) {
@@ -60,51 +64,52 @@ function createCoreApiHarness(appOverrides: any = {}) {
 
 describe("storage space admin API", function () {
 	it("requires admin read permission before reading storage-space aggregates", async () => {
-		let databaseCalled = false;
+		let storageSpaceCalled = false;
 		const {call} = createCoreApiHarness({
 			ms: {
-				database: {
+				storageSpace: {
 					getStorageSpaceOverview: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					getStorageSpaceTypeBreakdown: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					getStorageSpaceTopContents: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					getStorageSpaceTopFileCatalogItems: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					getStorageSpaceFileCatalogFolders: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					getStorageSpaceTopGroups: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					getStorageSpaceGroupPosts: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					getLatestStorageSpaceSnapshot: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					refreshStorageSpaceSnapshot: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 					queueStorageSpaceSnapshotRefresh: async () => {
-						databaseCalled = true;
+						storageSpaceCalled = true;
 						return {};
 					},
 				},
+				database: {},
 				communicator: {},
 				storage: {},
 			},
@@ -128,12 +133,12 @@ describe("storage space admin API", function () {
 			const response = await call(method, path, {user: {id: 7}});
 			assert.equal(response.body, 403);
 		}
-		assert.equal(databaseCalled, false);
+		assert.equal(storageSpaceCalled, false);
 	});
 
 	it("exposes storage-space aggregate helpers through admin read routes", async () => {
 		const permissionChecks = [];
-		const databaseCalls = [];
+		const storageSpaceCalls = [];
 		const query = {limit: "5", offset: "10"};
 		const responses = {
 			overview: {logicalContentBytes: 100, physicalContentBytes: 60},
@@ -149,48 +154,49 @@ describe("storage space admin API", function () {
 		};
 		const {call} = createCoreApiHarness({
 			ms: {
-				database: {
+				storageSpace: {
 					getStorageSpaceOverview: async () => {
-						databaseCalls.push(["overview"]);
+						storageSpaceCalls.push(["overview"]);
 						return responses.overview;
 					},
 					getStorageSpaceTypeBreakdown: async (listParams) => {
-						databaseCalls.push(["typeBreakdown", listParams]);
+						storageSpaceCalls.push(["typeBreakdown", listParams]);
 						return responses.typeBreakdown;
 					},
 					getStorageSpaceTopContents: async (listParams) => {
-						databaseCalls.push(["topContents", listParams]);
+						storageSpaceCalls.push(["topContents", listParams]);
 						return responses.topContents;
 					},
 					getStorageSpaceTopFileCatalogItems: async (listParams) => {
-						databaseCalls.push(["topFileCatalogItems", listParams]);
+						storageSpaceCalls.push(["topFileCatalogItems", listParams]);
 						return responses.topFileCatalogItems;
 					},
 					getStorageSpaceFileCatalogFolders: async (listParams) => {
-						databaseCalls.push(["fileCatalogFolders", listParams]);
+						storageSpaceCalls.push(["fileCatalogFolders", listParams]);
 						return responses.fileCatalogFolders;
 					},
 					getStorageSpaceTopGroups: async (listParams) => {
-						databaseCalls.push(["topGroups", listParams]);
+						storageSpaceCalls.push(["topGroups", listParams]);
 						return responses.topGroups;
 					},
 					getStorageSpaceGroupPosts: async (listParams) => {
-						databaseCalls.push(["groupPosts", listParams]);
+						storageSpaceCalls.push(["groupPosts", listParams]);
 						return responses.groupPosts;
 					},
 					getLatestStorageSpaceSnapshot: async () => {
-						databaseCalls.push(["snapshot"]);
+						storageSpaceCalls.push(["snapshot"]);
 						return responses.snapshot;
 					},
 					refreshStorageSpaceSnapshot: async (...args) => {
-						databaseCalls.push(["refreshSnapshot", ...args]);
+						storageSpaceCalls.push(["refreshSnapshot", ...args]);
 						return responses.refreshedSnapshot;
 					},
 					queueStorageSpaceSnapshotRefresh: async (...args) => {
-						databaseCalls.push(["queueSnapshot", ...args]);
+						storageSpaceCalls.push(["queueSnapshot", ...args]);
 						return responses.queuedSnapshot;
 					},
 				},
+				database: {},
 				communicator: {},
 				storage: {},
 			},
@@ -223,7 +229,7 @@ describe("storage space admin API", function () {
 			[7, CorePermissionName.AdminRead],
 			[7, CorePermissionName.AdminRead],
 		]);
-		assert.deepEqual(databaseCalls, [
+		assert.deepEqual(storageSpaceCalls, [
 			["overview"],
 			["typeBreakdown", query],
 			["topContents", query],
