+++ test/storageSpaceUsage.test.ts
@@ -36,7 +36,7 @@ describe("storage space usage", function () {
 	});
 
 	it("separates logical content bytes from deduplicated physical bytes", async () => {
-		const before = await app.ms.database.getStorageSpaceOverview();
+		const before = await app.ms.storageSpace.getStorageSpaceOverview();
 		const firstUser = await app.registerUser({
 			email: 'user@user.com',
 			name: 'user',
@@ -73,7 +73,7 @@ describe("storage space usage", function () {
 			status: PostStatus.Published
 		});
 
-		const after = await app.ms.database.getStorageSpaceOverview();
+		const after = await app.ms.storageSpace.getStorageSpaceOverview();
 		const sharedSize = Number(sharedContent.size);
 		const duplicateSize = Number(duplicateContent.size);
 		const uniqueSize = Number(uniqueContent.size);
@@ -85,22 +85,22 @@ describe("storage space usage", function () {
 		assert.equal(after.groupPostsLogicalBytes - before.groupPostsLogicalBytes, sharedSize);
 		assert.equal(after.fileCatalogLogicalBytes > before.fileCatalogLogicalBytes, true);
 
-		const typeBreakdown = await app.ms.database.getStorageSpaceTypeBreakdown({limit: 10});
+		const typeBreakdown = await app.ms.storageSpace.getStorageSpaceTypeBreakdown({limit: 10});
 		const textType = typeBreakdown.find(row => row.mimeType === 'text/plain');
 		assert.equal(!!textType, true);
 		assert.equal(textType?.logicalBytes, sharedSize + duplicateSize + uniqueSize);
 		assert.equal(textType?.physicalBytes, sharedSize + uniqueSize);
 		assert.equal(textType?.storageObjectsCount, 2);
 
-		const topContents = await app.ms.database.getStorageSpaceTopContents({limit: 5});
+		const topContents = await app.ms.storageSpace.getStorageSpaceTopContents({limit: 5});
 		assert.equal(topContents.some(row => row.id === uniqueContent.id), true);
 		assert.equal(topContents.every(row => Number.isFinite(row.size)), true);
 
-		const topCatalogItems = await app.ms.database.getStorageSpaceTopFileCatalogItems({limit: 5});
+		const topCatalogItems = await app.ms.storageSpace.getStorageSpaceTopFileCatalogItems({limit: 5});
 		assert.equal(topCatalogItems.some(row => row.contentId === sharedContent.id), true);
 		assert.equal(topCatalogItems.every(row => row.type === 'file'), true);
 
-		const rootFolders = await app.ms.database.getStorageSpaceFileCatalogFolders({limit: 5});
+		const rootFolders = await app.ms.storageSpace.getStorageSpaceFileCatalogFolders({limit: 5});
 		const usageFolder = rootFolders.find(row => row.name === 'usage');
 		assert.equal(!!usageFolder, true);
 		assert.equal(usageFolder?.logicalBytes, sharedSize + uniqueSize);
@@ -109,7 +109,7 @@ describe("storage space usage", function () {
 		assert.equal(usageFolder?.directFilesCount, 1);
 		assert.equal(usageFolder?.childFoldersCount, 1);
 
-		const childFolders = await app.ms.database.getStorageSpaceFileCatalogFolders({limit: 5, parentItemId: usageFolder.id});
+		const childFolders = await app.ms.storageSpace.getStorageSpaceFileCatalogFolders({limit: 5, parentItemId: usageFolder.id});
 		const deepFolder = childFolders.find(row => row.name === 'deep');
 		assert.equal(!!deepFolder, true);
 		assert.equal(deepFolder?.parentItemId, usageFolder?.id);
@@ -117,12 +117,12 @@ describe("storage space usage", function () {
 		assert.equal(deepFolder?.physicalBytes, uniqueSize);
 		assert.equal(deepFolder?.filesCount, 1);
 
-		const topGroups = await app.ms.database.getStorageSpaceTopGroups({limit: 5});
+		const topGroups = await app.ms.storageSpace.getStorageSpaceTopGroups({limit: 5});
 		const usageGroup = topGroups.find(row => row.id === group.id);
 		assert.equal(!!usageGroup, true);
 		assert.equal(usageGroup?.size, sharedSize);
 
-		const groupPosts = await app.ms.database.getStorageSpaceGroupPosts({limit: 5, groupId: group.id});
+		const groupPosts = await app.ms.storageSpace.getStorageSpaceGroupPosts({limit: 5, groupId: group.id});
 		const usagePost = groupPosts.find(row => row.id === sharedPost.id);
 		assert.equal(!!usagePost, true);
 		assert.equal(usagePost?.groupId, group.id);
@@ -132,8 +132,8 @@ describe("storage space usage", function () {
 		assert.equal(usagePost?.attachmentsCount, 1);
 		assert.equal(usagePost?.storageObjectsCount, 1);
 
-		assert.equal(await app.ms.database.getLatestStorageSpaceSnapshot(), null);
-		const snapshot = await app.ms.database.refreshStorageSpaceSnapshot(firstUser.id, {limit: 2, offset: 99});
+		assert.equal(await app.ms.storageSpace.getLatestStorageSpaceSnapshot(), null);
+		const snapshot = await app.ms.storageSpace.refreshStorageSpaceSnapshot(firstUser.id, {limit: 2, offset: 99});
 		assert.equal(snapshot.userId, firstUser.id);
 		assert.equal(snapshot.listLimit, 2);
 		assert.equal(snapshot.data.overview.logicalContentBytes, after.logicalContentBytes);
@@ -144,22 +144,22 @@ describe("storage space usage", function () {
 		assert.equal(snapshot.data.topGroups.length <= 2, true);
 		assert.equal(snapshot.data.groupPosts.length <= 2, true);
 
-		const latestSnapshot = await app.ms.database.getLatestStorageSpaceSnapshot();
+		const latestSnapshot = await app.ms.storageSpace.getLatestStorageSpaceSnapshot();
 		assert.equal(latestSnapshot.id, snapshot.id);
 		assert.deepEqual(latestSnapshot.data.overview, snapshot.data.overview);
 
-		const queuedRefresh = await app.ms.database.queueStorageSpaceSnapshotRefresh(firstUser.id, null, {limit: 1, offset: 99}, {process: false});
+		const queuedRefresh = await app.ms.storageSpace.queueStorageSpaceSnapshotRefresh(firstUser.id, null, {limit: 1, offset: 99}, {process: false});
 		assert.equal(queuedRefresh.module, "storage-space-snapshot");
 		assert.equal(queuedRefresh.isWaiting, true);
 
-		assert.deepEqual(await app.ms.database.processStorageSpaceSnapshotRefreshQueue({limit: 1}), {processed: 1});
+		assert.deepEqual(await app.ms.storageSpace.processStorageSpaceSnapshotRefreshQueue({limit: 1}), {processed: 1});
 		const processedQueue = await app.ms.asyncOperation.getUserOperationQueue(firstUser.id, queuedRefresh.id);
 		assert.equal(processedQueue.isWaiting, false);
 		assert.equal(processedQueue.asyncOperation.inProcess, false);
 		assert.equal(processedQueue.asyncOperation.percent, 100);
 
 		const queuedRefreshOutput = JSON.parse(processedQueue.asyncOperation.output);
-		const queuedSnapshot = await app.ms.database.getLatestStorageSpaceSnapshot();
+		const queuedSnapshot = await app.ms.storageSpace.getLatestStorageSpaceSnapshot();
 		assert.equal(queuedRefreshOutput.snapshotId, queuedSnapshot.id);
 		assert.equal(queuedSnapshot.listLimit, 1);
 		assert.equal(queuedSnapshot.data.topContents.length <= 1, true);
